V1
This is just the first version of the project. In the second version, the communication will be over HTTPS (port 443), the drawer will have position sensors which will make calibration (different timing and speed of motors) unnecessary. It will also work fine even if the nylon string gets stretched over time.

The electronics driving the painting will get some extra protection and speed control as well as signal encryption.